+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Plugin command labels (Removed `Tabs:` prefix)
 - Breadcrumbs and navigation icon styles
 
+### Removed
+
+- User option to specify minimum width of a single breadcrumb (`Minimum breadcrumb width (px)`)
+
 ## [1.2.1] - 2021-02-23
 
 ### Fixed